copyAsset: check if file exists before copying

+++ docco.litcoffee
@@ -84,7 +84,7 @@ out in an HTML template.
 
         callback or= (error) -> throw error if error
         copyAsset  = (file, callback) ->
-          fs.copy file, path.join(config.output, path.basename(file)), callback
+          if fs.existsSync file then fs.copy file, path.join(config.output, path.basename(file)), callback
         complete   = ->
           copyAsset config.css, (error) ->
             if error then callback error
